Update cloud9 install script
The script failed to install imagemagick and friends, but still exited
successfully and told me the application was running. My install was also
missing bower, which the install script assumes is already installed. This
commit fixes the install and makes sure that, if any of the install steps
fail, the whole script fails, so that users understand that an error
occurred.